fix(hooks): fold in the batch-4 review on #684
Three things the reviewers asked for, none of them a change of behaviour.

Dockerfile.dockerignore's header described the build context as the
workspace root, which is what this branch stopped it being. The patterns
still work and the `-f`-adjacent lookup was verified by hand, but the
explanation of why was left pointing at the old shape. It says what the
context actually is now: a scratch directory holding exactly the two
staged trees.

repo_root(), read() and generated_pre_push() move to tests/common/hooks.rs.
The heredoc marker string is the fragile part of the third one, and the
cost-control guards on top of this branch carry a second verbatim copy of
it, so a quoting change in install-hooks.sh would have needed the same edit
in two places and failed unhelpfully in whichever one did not get it. One
copy now, with the marker as a named constant.

run-tests.sh resolves its paths with `pwd -P`. It was printing logical
paths while staging physical ones, so a symlinked workspace made the
sibling-layout cell red against a layout that works fine.
